backend/pyproject.toml
[project]
name = "silverfish-backend"
version = "0.1.0"
description = "衣鱼 (Silverfish) - 小说人物关系解析系统"
requires-python = ">=3.10,<3.12"
license = { text = "AGPL-3.0" }
authors = [
    { name = "Silverfish Team" }
]

dependencies = [
    # 核心框架
    "flask>=3.0.0",
    "flask-cors>=6.0.0",
    
    # LLM 相关
    "openai>=1.0.0",
    
    # OASIS 社交媒体模拟
    "camel-oasis==0.2.5",
    "camel-ai==0.2.78",
    
    # 文件处理
    "PyMuPDF>=1.24.0",
    
    # 工具库
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pipreqs>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]
